requirements.txt
# Database support
psycopg2-binary==2.9.7

# No other external dependencies required
# The application uses Python standard library modules
